Create fatDamage.py
This is a python file that does the damage of a character I made, his whole thing is that he is really fat and rolls into enemies. Change the sp to the speed he is moving and the w to his current weight. I think my equation is more balanced but I tried matching it to dnd and real life physics as well